// src/app/models/item.ts
export class Item {
    id: number;
    name: string;
    idType: number;
    typeName: string;
    qtyOnHand: number;
    unitPrice: number;
    idClass: number;
    className: string;
    idCategory: number;
    categoryName: string;
    idUnit: number;
    unitName: string;
    inactive: boolean;
    sku: string;
    upc: string;
    defaultLocation: number;
    defaultLocationName: string;
  }